requirements.txt
imap-tools==1.11.0
python-dotenv==1.1.1
